#25 complete

+++ src/main/java/by/ladyka/club/dto/tikets/EventTicketsReportDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class EventTicketsReportDto {
-	private int danceCount;
+	private Long danceCount;
 	private int tablePlacesCount;
 }
